+++ src/extension_instance.mli
@@ -2,13 +2,15 @@ open Import
 
 type t
 
-val make : Sandbox.t -> (t, string) result Promise.t
+val make : Sandbox.t -> t Promise.t
 
 val sandbox : t -> Sandbox.t
 
-val language_client : t -> LanguageClient.t
+val language_client : t -> LanguageClient.t option
 
-val ocaml_lsp : t -> Ocaml_lsp.t
+val ocaml_lsp : t -> Ocaml_lsp.t option
+
+val lsp_client : t -> (LanguageClient.t * Ocaml_lsp.t) option
 
 val update_on_new_sandbox : t -> Sandbox.t -> (unit, string) result Promise.t