+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Improve highlighting of type parameters and `module type of` (#461)
 - Fix highlighting of escaped character literals (#467)
 - Fix highlighting of comments that contain strings with escaped quotes (#469)
+- Initialize extension even if language server fails to start (#471)
 
 ## 1.5.0